moved textStyles
